Update of 400 Bad Request

+++ WebPryton/Models/Account.cs
@@ -9,7 +9,7 @@ namespace WebPryton.Models
     {
         public string UserId { get; set; }
         public string Login { get; set; }
-        public int Password { get; set; }
+        public long Password { get; set; }
         public string Status { get; set; }
     }
 }
